+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `n_lookbacks` (which should have been called `n_baselines`) for computing confirmation within the SAS
    - Removed overly complicated accounting for in SAS confirmation
    - Only confirmation now is either by passing a previous product OR passing a directory of products
-- Constant for `BASE_DATE_FOR_CONFIRMATION`.
+- Constant for `BASE_DATE_FOR_CONFIRMATION` and removed from algo config.
 - Extra validation for confirmation packaging
+- Support for `immediate_lookback` - to add back will need to add logic for calculations of lookbacks, etc.
 
 ## Changed
 - Multiwindow strategy is now the default in both python API and CLI
@@ -30,10 +31,14 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Variables for confirmation processing (use snake case wherever possible)
 - Now have an output product without confirmation and a product with confirmation (for provenance)
 - De-couple confirmation and packaging in workflow logic
-- Confidence is now float32 (not int16) - this needs to happen because our metrics is a float32.
-- Updated golden dataset to reflect new changes.
+- Confidence is now float32 (not int16) - this needs to happen because the dist metrics is float32 so casting to integer looses information.
+- Updated golden dataset to reflect new changes in this major release.
 - `DIST_CMAP` is not `DIST_STATUS_CMAP`.
 - To align with dist-hls, we have changed `moderate_confidence_threshold` and `high_confidence_threshold` to now be `low_confidence_alert_threshold` and `high_confidence_alert_threshold`.
+- Moved browse imagery generation into workflows.
+- Consistent multiprocessing within dist-s1 (ensures consistent useage of `spawn` child processes).
+- Handling of model context length to provide maximum number of pre-images (can still be overwritten)
+- Removal of defaults using lists! https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments
 
 ## Added
 - Confirmation python API and CLI
@@ -55,12 +60,14 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Confirmation CLI
   - Ensuring additional profile keys  for COG in default GTiff are not present.
 - Access of Confirmation of Products in main library (i.e. `from dist_s1 import run_sequential_confirmation_of_dist_products_workflow`)
+- Use specified model context length from specified model to calculate `max_pre_imgs_per_burst_mw` and `delta_lookback_days_mw` if the latter two are unspecified
 
 ## Fixed
 - Sequential Confirmation Workflow and downstream functions
 - All the tests with the updates above.
 - Correct loading of algo_config.yml in `prep` steps.
   - Only `from_yml` loads the `algo_config` correctly, but when it is assigned in the `prep` workflows (i.e. the attribute `algo_config` is assigned, this yml is not correctly loaded.
+- Correctly handle `model_source` as only string value with allowed string values from `distmetrics` or `external`.
 
 
 ## [1.0.1] - 2025-06-05